pyproject.toml
[project]
name = "dubsmith"
version = "0.6.0"
description = "Auto-fetch missing language dubs from Crunchyroll and mux them into a Sonarr-managed library"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "numpy>=1.26",
  "scipy>=1.13",
  "PyYAML>=6.0",
  "click>=8.1",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "APScheduler>=3.10",
  "Jinja2>=3.1",
  "python-multipart>=0.0.9",
  "itsdangerous>=2.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
dubsmith = "src.main:cli"
plex-dub = "src.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --tb=short"
